backend/requirements.txt
azure-functions>=1.15.0
azure-cosmos>=4.5.1
azure-keyvault-secrets>=4.7.0
azure-identity>=1.14.0
requests>=2.31.0